pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "hijri-converter>=2.3.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pdfkit>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "python-bidi>=0.6.6",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "xlsxwriter>=3.2.2",
]
